moving core to share extension
